pyproject.toml
[project]
name = "dependapy"
version = "0.1.1"
description = "A custom dependabot alternative for Python projects"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
    {name = "stefanposs", email = "hello@stefanposs.com"},
]
keywords = ["dependencies", "update", "dependabot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Version Control",
]

dependencies = [
    "requests==2.32.4",
    "packaging==25.0",
    "tomli==2.2.1",
    "PyGithub==2.6.1",
    "ruff==0.11.13",
]


[dependency-groups]
dev = [
    "pytest==7.4.0",
    "pytest-mock==3.11.0",
    "ruff==0.11.13",
    "pyright==1.1.401",
    "bandit==1.7.5",
    "pytest-cov==4.1.0",
    "twine>=6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dependapy"]

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
dependapy = "dependapy.main:main"

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101", "B404", "B603", "B607"]

[tool.pyright]
include = ["dependapy"]
exclude = ["tests"]
typeCheckingMode = "basic"
stubPath = "typings"
reportMissingImports = true
